requirements.txt
# Core UI and Processing
customtkinter
ffmpeg-python
numpy

# Speech Recognition
faster-whisper
whisper
whisperx

# Multimodal AI Models
torch
torchaudio
transformers
accelerate

# Video/Audio Processing
librosa
opencv-python
pillow
google-generativeai

# Additional AI Dependencies
datasets
requests
